Improved support for viewports on tilesort -- they now interact correctly with the bucketer.
Made it so the non-presence of a mothership is no longer fatal -- it's up
to the client to recover if it can.  The render spu will just use its
defaults.  This makes it possible to mess around with the SPU tester
without running a mothership every time.

Added support for setting a "fake" window size for the tilesort SPU in
the config file.  This is useful if the SPU is loaded directly without
a window backing it, so we know what to do with a Viewport() call.