Implemented byte-swapping.  This has been tested from IRIX->WINDOWS and Linux->Windows (to make sure everything still ran OK on the same endiannesses).
I tested it with the pack SPU and the Tilesort SPU, and everything was
hunky dory with fonttest and atlantis (running over a DSL line, got
0.6 FPS on atlantis!)

Also, fixed a bug where parallel API commands were not being broadcast
by the tilesort SPU (whoops!)

All connections are now brokered through the mothership, which is how
the client figures out the server's endianness.

All the swapping work is done by the client, including the swapping
of the return values of round-trip glGet() calls in the pack SPU.  That
was the biggest kludge of this whole thing, but it appears to work OK.

None of this code has been tested on Windows (I've been using an out-of-date
server on Windows that just expects to get everything in its native
endianness, which is exactly what's happening).