Turn on the SO_REUSEADDR option for mothership and system sockets so that life sucks less on Linux (wonder why Windows never has the "address already in use" problem?)
Also, make the default for the render SPU's force_direct option be zero, since
it's a little ham-fisted to INSIST on a direct rendering visual if we tried
and failed.  Possibly we should check the driver version and insist if it's
a driver that's known not to work with indirect rendering.