Started a simple framework for the readback SPU.  Currently works OK in the very simple configuration given in the sample configuration script.
Also, extended the semantics of glBarrierCreate so that a count of "0" means
"make a barrier with a size equal to the number of connected clients" once
the command reaches the server.  This is cleaner than trying to do some
serious analysis on a complex graph.  It's quite uncommon to want to do partial
barriers, and if that comes up we can always extend the semantics further to
have negative numbers mean "all the connected clients except for $n$ of them"
or something.