Fixed a number of bugs for Sean.  Amazing how productive SIGGRAPH can make one.
The length table in the pincher was incomplete, causing crashes.
The pincher couldn't handle multitexturing, causing errors.

The pincher was putting MultiTexCoord calls on the wire unnecessarily
because it was a) ignoring the limits structure, and b) not optimizing
a glMultiTexCoord( GL_TEXTURE0_ARB, ... ) into a glTexCoord( ... ).

Because the SGI's don't HAVE Multitexturing, this was causing an IRIX
server to complain and exit whenever a pinch happened.  Of course, since
an IRIX box should always report that it has one texture unit, this
should never happen again.

I tested these changes against MeshTV, and what did not render before now
renders perfectly.  I built the code on Linux.

This represents some pretty severe changes to the system, including
some shuffling of protocol numbers, so make sure you do a 'make' from
the toplevel directory.  Heck, maybe even a paranoid 'make clean' is in
order after this one.

	-Greg