Added the ability to use a file as a "network" for dump/playback.
Added file functionality to the mothership -- you can now create
a "server" node with a file "client" without having to have an actual
node somewhere.  This means that it now makes sense to have a graph
with a single server node, which is the "normal" way to do playback.

Added a SPU that intercepts all glGets and handles them, passing
everything else through.  You can use this if you want to create a
trace of an application that's going to query things about the
OpenGL machine, like the max texture size, etc.  It's called
"simplequery".