Changes to get file networking working again.  Some of these fixes might be considered temporary and should probably be reviewed by someone.
1)  There was a bug in server_config where we were identifying ourselves
as a client SPU twice.  Fix that.

2)  Don't try to return values to the client in server_retval if the
network connection is a file.

3)  Make the mothership say something other than "Client blew up" because
that's causing people to think that the error is in the mothership window.
Now say something more polite that will make it obvious that the client
did something silly.

4)  Don't try to receive values from the server when packing CreateContext
to a file.

5)  Artificially number the contexts that packspu_CreateContext generates
so they match the server's expected magic numbers.  This is supremely gross
and there is a large comment in the file to that effect.