Moved crPackInit from cr_packfunctions.h to cr_pack.h.  What was it doing there in the first place?
Changed the welcome message from the packer to be more useful.

Added a small test program to measure the raw performance of the packer.
Currently, with things building optimized, we seem to be able to pack about
20MVerts/sec on Windows 2000, 800 MhZ PIII, RDRAM, which is nice because
it's the same number I measured a year ago.  I like it when experiments
repeat.

Building the pack library in debug mode seems to kill about 25% of the
performance, which makes sense.