Got rid of the silly cube code in the render SPU.
Made the render SPU window not always on top on Windows.  I don't know if this
was done on purpose, but it makes debugging on Windows nearly impossible.

Added some random number generation stuff to the utility library.

Added a fun SPU that sort-of simulates a wet camera.  It bears little relation
to physics, but it is a pretty neat effect.  It currently has an off-by-one
error somewhere that causes the edges of the images to be wrong, and some piece
of state is not completely reset (the guns in quake 3 are drawn wrong, probably
I'm clobbering one of their textures because I'm bad and didn't use
glGenTextures), but it's still fun to play with.