Fix maximum viewport size in the tilesort SPU.
Here I've created a new GL_SET_MAX_VIEWPORT_CR parameter to pass through
with ChromiumParametervCR(). We issue a call to each server node to access
it's GL_MAX_VIEWPORT_DIMS and total up the results from each. Finally
issuing the ChromiumParametervCR() to setup the new maximum viewport
in the servers.

This is all done in the tilesortspu_init.c function before any contexts
are created.