Fix glFinish and use round-trip crPackWriteback call. Remove the same obsolete code from glCopyPixels
Also, ensure that a packer context is always set. This problem shows
up with glean that calls DestroyContext and resets the packer context
to NULL. We never do that now, but ensure we've got a current packer
context via CreateContext and MakeCurrent, as we'll always need
a valid packer context to actually send any data.