More fixes to the glX visual handling.
Now, we call StubInit() in glXGetConfig and glXChooseVisual. The problem
was that stub.desiredVisual was being set in these two functions and
then reset when StubInit() was called, resetting the requested
attributes. Doing StubInit() in these functions removes that problem.

Also, we only break up the visual information when no desiredflags have
already been set.