Added the "only_swap_once" option to the server, which makes it listen to swap and (color) clears from only one client.  This will simplify the design of parallel applications, since they don't have to conditionally submit clear and swap commands.
Note that there's still no auto_barrier option for the tilesort SPU; it's
coming next.

Also made more progress on the parallel modelview application, which motivated
some of these changes.

Finally, cleaned up some naming conventions in the render SPU to make it
more clear which functions are generic and which are system-specific.  Added
more debugging output when selecting visuals and creating windows fail, as
the current failure modes were a bit mysterious.