Export Performance data to applications for complete interaction.
This allows any application to reset, gather, change etc the
performance data, by retrieving a pointer to that information through
glGetChromiumParametervCR().

This is much easier than implementing RESET calls for various vertex
data and is much more flexible.

The structure for the performance data is moved to cr_applications.h