Add a new crNetDump() to give back the connection blocks to the calling function.
This allows the perfSPU (at cleanup time) to dump the total bytes
sent and received on each outstanding connection. The dumping of
this data is primitive and needs formatting.

This doesn't catch mothership connections as they're closed once finished
with, but catches the raw data being sent between the crservers using
util/net.c counters.