Add Feedback/Selection support (feedbackspu)
There's a little bit of duplicate code in the state_feedback.c file, but
I've left it like that for clarity, as I doubt very much that this file
will change that much.

Passes conformance tests too.

There's no configuration to be done with this SPU, just insert it into
the SPU chain and it'll do the rest.