Added semaphores to deal with rasterpos issues, this should help some of the flashing.
Added initial support for clipping to object.  This is a little messy at the moment, I'll clean it up soon.  The GLHint hack is back in temporarily for testing a few apps, it will be removed once all of this is known to work.

Lots of clean up, should be a little faster now.

Still need to deal with resizable windows once we get CRUT...  We really need CRUT for sort last...