Last commit should have included all 4 files, crutclientapi and crutapi source and header files.
Modified crutMainLoop to discard mouse motion and passive motion events when last event processed was the same as the next event in the queue.  This way, the application processes motion events as fast as it possibly can, and avoids processing unnecessary events.  If the application gets to the end of the queue, the next event to enter is always processed, so effectively the events sent over the network are sampled at the rate at which the application can handle them.  Interaction is now very fluid for applications such as a model viewer.

Also fixed extern "C" bug with header files, and changed crutSendXEvent so there is a function that handles all event dispatching.