Check-in on behalf of Bob Ellison:
These changes remove a race condition during autostart.

Formerly, the main application would spawn the mothership,
and sleep for a second while the mothership spawned the crservers.
Sometimes this was enough time; sometimes it was not.

Now, the main application tells the mothership that it would like
to be signalled when the crservers have been spawned; it then
sleeps until it receives this signal (or until 60 seconds have passed,
when it gives a warning but then continues).

This code should have no effect on Windows implementations; the
Python docs say that Windows Python doesn't support the os.kill()
routine, required to send a signal, so the Windows implementation
works as before.