These changes support threaded autostart, and dynamic node markers.
Threaded Autostart: the mothership now accepts options that control
how nodes are autostarted.  By default (which matches previous behavior),
a single thread in the mothership is started, which starts all the
autostart nodes sequentially.

Now you can specify that the mothership create multiple threads,
and even trees of threads, to spawn autostart nodes.

This functionality is tentative; it is provided for examination and
benchmarking, to see if it's faster on large clusters than a
sequential startup.  It may be removed if it shows no particular utility.

--------------------------------------------------------

Dynamic node markers: Instead of a host name, a Chromium
config file may now specify a host marker, in the form "#" + name
(e.g. "#1", "#2", "#server1", etc.), for almost any node.

The first node of the appropriate type (i.e. a crappfaker won't
be mistaken for a crserver) that connects to the mothership
will define an unresolved marker, and will be used forthwith.

Note that if any markers exist in the Chromium graph defined
for the mothership, Chromium won't let the main application
continue until all the unspecified nodes have connected (because
the main application will need a list of all servers, with all
names resolved); this can effectively slow down startup.

Also note that this feature is not likely to work on Windows,
because the mechanism used by the mothership to signal the
main application to continue (a UNIX signal, via the Python os.kill()
routine) doesn't appear to exist in Windows Python.