Dynamic Host Marker Revisions
These are modifications, suggested by Joel Welling, to allow
dynamic host markers to work better on widely distributed networks.

They introduce the concept of a "dynamic constraint" - a constraint
on which hostnames are allowed to match a particular dynamic marker.

As before, to introduce a dynamic marker, specify a string
starting with "#" as the node name:

	server = CRNetworkNode("#server1")

To specify constraints, add the "dynamic_constraints" attribute,
with an array of one or more constraint specifications:

	server.Conf('dynamic_constraints', [ 'test:predicate', ...])

Each "test" is a test name; the associated predicate is test-specific,
and dictates somehow what the hostname is supposed to match.  Only
one test is initially defined; it is called "name", and the predicate
is a regular expression that a hostname will have to match:

	server = CRNetworkNode("#server1")
	server.Conf('dynamic_constraints', [ 'name:iamvis[0-9]*\\.psc\\.edu'])

means that any prospective hostname seeking to match the dynamic
marker "#server1" will have to be able to match the specified regular
expression to be considered.