A generalization of the FPS spu; behavior that used to be controlled by hard-coded constants is now managed by configurable parameters.
    fpsspu.Conf('report_frames', 500)

requires the SPU to give a frames-per-second report at least once
every 500 frames.  (The default is 10, the old hard-coded value.)

    fpsspu.Conf('report_seconds', 3.5)

requires the SPU to give a frames-per-second report if the interval
between reports gets above 3.5 seconds.  (The default is 1.0 second,
the old hard-coded value).

If both are provided, either can trigger a report.  If either is set
to 0, that trigger is disabled.  To get a report every 1000 frames,
for example, use
    fpsspu.Conf('report_frames', 1000)
    fpsspu.Conf('report_seconds', 0.0)

One more value is provided:
    fpsspu.Conf('report_at_end', 1)

This is a new feature; it provides an average FPS across the whole
run of the program, from the first SwapBuffers to shutdown.  (Note
that it doesn't count time from startup, just from the first SwapBuffers,
in order to eliminate startup overhead from the calculation.)
By default, this parameter is 0 (which disables the end report).