Added the Matte SPU, which allows a window to be divided into a a "render area" (essentially a subwindow inside the render window) and a "matte area" (the border around that subwindow).
This has been used in a tiled configuration involving a custom hardware
compositor that had specific requirements for the display:
    - only a small subarea (up to 512x512) of each tiled display would
      be used for compositing
    - The placement of that subarea was critical
    - The rest of the display had to be totally blank

The configuration mentioned above uses the Render SPU to render to
each display; every display is configured as a full-screen window,
but with only a 512x512 subtile.  The Matte SPU is used to place that
subtile in the correct position on the screen, while also clearing
the rest of the screen.