Added the "grabber" SPU.
This passthrough SPU is designed to intercept some basic operations and
pass information back to the mothership through crMothershipSetParam().
This information can then be requested by other Chromium nodes.

Right now, the only supported intercepted operation is MakeCurrent(); if
the SPU is configured with "current_window_attribute_name" set, the
window ID will be passed back to the mothership with the given name.

This can be used, for example, by a configuration that needs a render
node to render into an external window (i.e., something like the CRUT
utilities offer).  In fact, if the Render SPU is configured with
"render_to_crut_window" set, and the Grabber SPU is configured with
"current_window_attribute_name" set to "crut_drawable", the external
window grabbed by the Grabber SPU can be made to operate like a CRUT
window.