Added "read_size" and "draw_offset" configuration parameters, and upgraded "draw_bbox_outlines" to a true configuration parameter.
In a cluster running a parallel application but with each application
node rendering only a portion of the display (that will ultimately
be read back and reassembled), you can end up with a Chromium configuration
without node servers, and no way to get tiling information necessary
for reassembly.

The "draw_offset" parameter allows a particular Readback SPU instance
to know where it should be offsetting its DrawPixels() command, allowing
the individual application nodes to understand enough about tiling to
function.

The "read_size" is useful if the application nodes are all set to
"render_to_app_window"; in this case, the size used by ReadPixels() is
(mistakenly?) taken to be the size of the hidden Render window (256x256,
usually).  Setting "read_size" allows the size to be more precisely
controlled.