Fix a problem with visual selection found in the GeoProbe application, which used one X display connection to query about the visuals available on a screen, but a second connection to actually choose the GLX visual (probably due to the workings of a toolkit).
Chromium associates visual attributes with a tuplet of
(display pointer, screen, visual id), and did not recognize that the
second reference was to the same visual; it chose a default visual
type which did not include all the necessary components for the
application.

Now, if a visual is requested during glXCreateContext that has never
been seen before, its attributes are queried, to get a more correct
idea of its characteristics before creating the visual for use with
Chromium.

This fixes the behavior of the GeoProbe application.