Fixed a bug that sometimes caused lost connections in a crserver to corrupt connections that were still up, if "sdp" or "tcpip" connection types were used for a client.
These two connection types can, during normal use, detect that a
connection has gone dead; if that happens, they disconnect the connection
and set the conn->type field to CR_NO_CONNECTION to indicate the lost
connection.

Later, the crserver detects the lost client, and explicitly calls
conn->Disconnect(), which attempts to disconnect the connection a second
time.  This is mostly benign, except if the connection module uses the
conn->index field to keep track of the connection's entry in a private
array.  In this case, the conn->index field, which is invalid since
the connection has already been disconnected, may still point to a
more recently created and valid entry.

The connection for the newly created entry is then disrupted.  The
connection itself is not lost, but the record of its socket file descriptor
is, so the module may no longer receive incoming data.  This can cause
apparent hangs and/or spins.

The fix affects both of the modules that use the CR_NO_CONNECTION
indicator as an internal flag; these modules will now detect when a
dead connection is being formally cleaned up, and will not corrupt their
internal data structures when this happens.