Fixed a very rare bug.  In searching for a client to service, the crserver goes through the list of clients, cleaning up after clients if their connections have gone dead.
The original code was written to avoid disposing of a client while it
still had messages to process; but instead of referring to the correct
client, it referred to "curClient".  This allowed the crserver to dispose
of a client that still had messages waiting, as long there was another
client that was the current client that had no messages waiting.

The new code fixes the reference.