Added new configuration options that allow signals to be used to insert markers into the output stream:
    printspu.Conf('marker_signal', 12)

configures the "print" SPU to add a marker message to its output stream when
it receives signal 12 (SIGUSR2).

    printspu.Conf('marker_text', "any string $(signal) $(date) $(time)")

gives the format of the message so added to the output stream.  The patterns
"$(signal)", "$(date)", and "$(time)" are replaced with the signal number,
date (as "YYYY-MM-DD") and time (as "HH:MM:SS", using a 24-hour clock),
respectively.

This capability is used during debugging, to help isolate in a logfile specific
events of interest.