Fix a problem with glReadPixels blowing up the crserver.
Looks like a shared library/linker problem.  The address of the global
cr_unpackData var was different in the unpacker and crserver code.  In
the later place, it was always NULL.
Fixed by adding a new crUnpackGetDataPointer() function to return the value
of cr_unpackData.  Call that instead of directly accessing the variable.